+++ arcane/src/arcane/utils/APReal.h
@@ -1,27 +1,1 @@
-﻿// -*- tab-width: 2; indent-tabs-mode: nil; coding: utf-8-with-signature -*-
-//-----------------------------------------------------------------------------
-// Copyright 2000-2022 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
-// See the top-level COPYRIGHT file for details.
-// SPDX-License-Identifier: Apache-2.0
-//-----------------------------------------------------------------------------
-/*---------------------------------------------------------------------------*/
-/* APReal.h                                                    (C) 2000-2018 */
-/*                                                                           */
-/* Réel en précision arbitraire.                                             */
-/*---------------------------------------------------------------------------*/
-#ifndef ARCANE_UTILS_APREAL_H
-#define ARCANE_UTILS_APREAL_H
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
-#include "arccore/base/APReal.h"
-namespace Arcane
-{
-using Arccore::APReal;
-}
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
-#endif
-
+﻿#include "arccore/base/APReal.h"
