+++ arcane/src/arcane/utils/ArrayIterator.h
@@ -1,30 +1,1 @@
-﻿// -*- tab-width: 2; indent-tabs-mode: nil; coding: utf-8-with-signature -*-
-//-----------------------------------------------------------------------------
-// Copyright 2000-2022 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
-// See the top-level COPYRIGHT file for details.
-// SPDX-License-Identifier: Apache-2.0
-//-----------------------------------------------------------------------------
-/*---------------------------------------------------------------------------*/
-/* ArrayIterator.h                                             (C) 2000-2018 */
-/*                                                                           */
-/* Itérateur sur les Array, ArrayView, ConstArrayView, ...                   */
-/*---------------------------------------------------------------------------*/
-#ifndef ARCANE_UTILS_ARRAYITERATOR_H
-#define ARCANE_UTILS_ARRAYITERATOR_H
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
-#include "arccore/base/ArrayIterator.h"
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
-namespace Arcane
-{
-using Arccore::ArrayIterator;
-} // End namespace Arcane
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
-#endif  
+﻿#include "arccore/base/ArrayIterator.h"
