+++ arcane/src/arcane/utils/BasicTranscoder.h
@@ -1,22 +1,2 @@
-﻿// -*- tab-width: 2; indent-tabs-mode: nil; coding: utf-8-with-signature -*-
-//-----------------------------------------------------------------------------
-// Copyright 2000-2022 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
-// See the top-level COPYRIGHT file for details.
-// SPDX-License-Identifier: Apache-2.0
-//-----------------------------------------------------------------------------
-/*---------------------------------------------------------------------------*/
-/* BasicTranscoder.h                                           (C) 2000-2022 */
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
-#ifndef ARCANE_UTILS_BASICTRANSCODER_H
-#define ARCANE_UTILS_BASICTRANSCODER_H
-
-#include "arccore/base/BasicTranscoder.h"
+﻿#include "arccore/base/BasicTranscoder.h"
 #include "arcane/utils/UtilsTypes.h"
-namespace Arcane
-{
-using Arccore::BasicTranscoder;
-}
-
-#endif
