+++ arcane/src/arcane/utils/Ref.h
@@ -1,41 +1,2 @@
-﻿// -*- tab-width: 2; indent-tabs-mode: nil; coding: utf-8-with-signature -*-
-//-----------------------------------------------------------------------------
-// Copyright 2000-2024 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
-// See the top-level COPYRIGHT file for details.
-// SPDX-License-Identifier: Apache-2.0
-//-----------------------------------------------------------------------------
-/*---------------------------------------------------------------------------*/
-/* Ref.h                                                       (C) 2000-2024 */
-/*                                                                           */
-/* Gestion des références sur une instance.                                  */
-/*---------------------------------------------------------------------------*/
-#ifndef ARCANE_UTILS_REF_H
-#define ARCANE_UTILS_REF_H
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
-#include "arccore/base/Ref.h"
+﻿#include "arccore/base/Ref.h"
 #include "arcane/utils/UtilsTypes.h"
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
-namespace Arcane
-{
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
-using Arccore::makeRef;
-using Arccore::makeRefFromInstance;
-using Arccore::createRef;
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
-}
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
-#endif
