+++ arccore/src/base/arccore/base/BaseTypes.h
@@ -1,11 +1,11 @@
 ﻿// -*- tab-width: 2; indent-tabs-mode: nil; coding: utf-8-with-signature -*-
 //-----------------------------------------------------------------------------
-// Copyright 2000-2024 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
+// Copyright 2000-2025 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: Apache-2.0
 //-----------------------------------------------------------------------------
 /*---------------------------------------------------------------------------*/
-/* BaseTypes.h                                                 (C) 2000-2024 */
+/* BaseTypes.h                                                 (C) 2000-2025 */
 /*                                                                           */
 /* Définition des types de la composante 'base' de Arccore.                  */
 /*---------------------------------------------------------------------------*/
@@ -33,8 +33,8 @@ namespace Arccore
  * \brief Déclarations des types de la composante 'base' de %Arccore.
  */
 
-template<typename T> class IterT;
-template<typename T> class ConstIterT;
+template <typename T> class IterT;
+template <typename T> class ConstIterT;
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
@@ -53,22 +53,24 @@ typedef float Single;
 //! Indique que la dimension d'un tableau est dynamique
 inline constexpr Int32 DynExtent = -1;
 
-template<typename T> class ConstArrayView;
-template<typename T> class ArrayView;
-template<typename T> class ConstArray2View;
-template<typename T> class Array2View;
-template<typename T> class ConstArray3View;
-template<typename T> class Array3View;
-template<typename T> class ConstArray4View;
-template<typename T> class Array4View;
-template<class DataType> class CoreArray;
+template <typename T> class ConstArrayView;
+template <typename T> class ArrayView;
+template <typename T> class ConstArray2View;
+template <typename T> class Array2View;
+template <typename T> class ConstArray3View;
+template <typename T> class Array3View;
+template <typename T> class ConstArray4View;
+template <typename T> class Array4View;
+template <class DataType> class CoreArray;
 template <typename T, typename SizeType, SizeType Extent = DynExtent, SizeType MinValue = 0> class SpanImpl;
 template <typename T, Int64 Extent = DynExtent, Int64 MinValue = 0> class Span;
 template <typename T, Int32 Extent = DynExtent, Int32 MinValue = 0> class SmallSpan;
 template <typename T, Int64 Extent = DynExtent, Int64 MinValue = 0> using LargeSpan = Span<T, Extent, MinValue>;
 template <typename T, typename SizeType, SizeType Extent1 = DynExtent, SizeType Extent2 = DynExtent> class Span2Impl;
-template<typename T, Int64 Extent1 = DynExtent, Int64 Extent2 = DynExtent> class Span2;
-template<typename T, Int32 Extent1 = DynExtent, Int32 Extent2 = DynExtent> class SmallSpan2;
+template <typename T, Int64 Extent1 = DynExtent, Int64 Extent2 = DynExtent> class Span2;
+template <typename T, Int32 Extent1 = DynExtent, Int32 Extent2 = DynExtent> class SmallSpan2;
+template <typename Iterator_>
+class ArrayIterator;
 
 class StringImpl;
 class String;
@@ -82,11 +84,11 @@ class Exception;
 class StackFrame;
 class FixedStackFrameArray;
 class IStackTraceService;
-template<typename T>
+template <typename T>
 class CheckedPointer;
-template<class T>
+template <class T>
 class ReferenceCounterAccessor;
-template<typename T>
+template <typename T>
 class ReferenceCounter;
 class ReferenceCounterImpl;
 class ArgumentException;
@@ -97,16 +99,32 @@ class NotImplementedException;
 class TimeoutException;
 enum class eBasicDataType : unsigned char;
 class IFunctor;
-template<typename ArgType>
+template <typename ArgType>
 class IFunctorWithArgumentT;
-template<typename ReturnType,typename Arg1,typename Arg2>
+template <typename ReturnType, typename Arg1, typename Arg2>
 class IFunctorWithArgAndReturn2;
-template<typename T>
+template <typename T>
 class FunctorT;
-template<typename ClassType,typename ArgType>
+template <typename ClassType, typename ArgType>
 class FunctorWithArgumentT;
-template<typename ArgType>
+template <typename ArgType>
 class StdFunctorWithArgumentT;
+class ReferenceCounterImpl;
+template <typename InstanceType, class T = void>
+struct RefTraits;
+template <typename InstanceType, int ImplTagId = RefTraits<InstanceType>::TagId>
+class Ref;
+
+namespace impl
+{
+  template <typename InstanceType>
+  class ReferenceCounterWrapper;
+}
+
+// Ces classes sont internes à Arccore/Arcane
+template <typename T>
+class ArrayRange;
+class BasicTranscoder;
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
@@ -202,6 +220,64 @@ typedef Span<const Integer> IntegerConstSpan;
 /*---------------------------------------------------------------------------*/
 
 } // End namespace Arccore
+
+namespace Arcane
+{
+using Arccore::Array2View;
+using Arccore::Array3View;
+using Arccore::Array4View;
+using Arccore::ArrayIterator;
+using Arccore::ArrayView;
+using Arccore::ConstArray2View;
+using Arccore::ConstArray3View;
+using Arccore::ConstArray4View;
+using Arccore::ConstArrayView;
+using Arccore::ConstIterT;
+using Arccore::CoreArray;
+using Arccore::eBasicDataType;
+using Arccore::IterT;
+using Arccore::SmallSpan;
+using Arccore::SmallSpan2;
+using Arccore::Span;
+using Arccore::Span2;
+using Arccore::Span2Impl;
+using Arccore::SpanImpl;
+using Arccore::StringImpl;
+
+using Arccore::TraceInfo;
+using Arccore::StackTrace;
+using Arccore::Exception;
+using Arccore::StackFrame;
+using Arccore::FixedStackFrameArray;
+using Arccore::IStackTraceService;
+using Arccore::CheckedPointer;
+using Arccore::ReferenceCounter;
+using Arccore::ReferenceCounterImpl;
+using Arccore::RefTraits;
+using Arccore::Ref;
+using Arccore::ArgumentException;
+using Arccore::IndexOutOfRangeException;
+using Arccore::FatalErrorException;
+using Arccore::NotSupportedException;
+using Arccore::NotImplementedException;
+using Arccore::TimeoutException;
+
+using Arccore::IFunctor;
+using Arccore::IFunctorWithArgumentT;
+using Arccore::IFunctorWithArgAndReturn2;
+using Arccore::FunctorT;
+using Arccore::FunctorWithArgumentT;
+using Arccore::StdFunctorWithArgumentT;
+
+using Arccore::Byte;
+using Arccore::SByte;
+using Arccore::Single;
+using Arccore::UChar;
+using Arccore::UInt16;
+
+// Ces classes sont internes à Arccore/Arcane
+using Arccore::BasicTranscoder;
+}
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
