+++ arccore/src/base/arccore/base/Ref.h
@@ -403,5 +403,23 @@ createRef(Args&&... args)
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
+namespace Arcane
+{
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+using Arccore::makeRef;
+using Arccore::makeRefFromInstance;
+using Arccore::createRef;
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+}
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
 #endif
