+++ arccore/src/concurrency/arccore/concurrency/ConcurrencyGlobal.h
@@ -1,11 +1,11 @@
 ﻿// -*- tab-width: 2; indent-tabs-mode: nil; coding: utf-8-with-signature -*-
 //-----------------------------------------------------------------------------
-// Copyright 2000-2024 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
+// Copyright 2000-2025 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: Apache-2.0
 //-----------------------------------------------------------------------------
 /*---------------------------------------------------------------------------*/
-/* ConcurrencyGlobal.h                                         (C) 2000-2024 */
+/* ConcurrencyGlobal.h                                         (C) 2000-2025 */
 /*                                                                           */
 /* Définitions globales de la composante 'Concurrency' de 'Arccore'.         */
 /*---------------------------------------------------------------------------*/
@@ -81,6 +81,30 @@ createNullThreadImplementation();
 /*---------------------------------------------------------------------------*/
 
 } // namespace Arccore::Concurrency
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+namespace Arcane
+{
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+using Arccore::IThreadImplementation;
+using Arccore::Mutex;
+using Arccore::SpinLock;
+using Arccore::GlobalMutex;
+using Arccore::IThreadBarrier;
+using Arccore::ThreadImpl;
+using Arccore::MutexImpl;
+using Arccore::NullThreadImplementation;
+using Arccore::NullThreadBarrier;
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+}
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
