+++ arccore/src/trace/arccore/trace/TraceGlobal.h
@@ -1,11 +1,11 @@
 ﻿// -*- tab-width: 2; indent-tabs-mode: nil; coding: utf-8-with-signature -*-
 //-----------------------------------------------------------------------------
-// Copyright 2000-2022 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
+// Copyright 2000-2025 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: Apache-2.0
 //-----------------------------------------------------------------------------
 /*---------------------------------------------------------------------------*/
-/* TraceGlobal.h                                               (C) 2000-2020 */
+/* TraceGlobal.h                                               (C) 2000-2025 */
 /*                                                                           */
 /* Définitions globales de la composante 'Trace' de 'Arccore'.               */
 /*---------------------------------------------------------------------------*/
@@ -64,6 +64,31 @@ namespace Trace
 
 ARCCORE_DECLARE_REFERENCE_COUNTED_CLASS(ITraceStream)
 ARCCORE_DECLARE_REFERENCE_COUNTED_CLASS(ITraceMng)
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+namespace Arcane
+{
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+using Arccore::ITraceStream;
+using Arccore::ITraceMng;
+using Arccore::TraceAccessor;
+using Arccore::TraceMessageClass; 
+using Arccore::TraceClassConfig;
+using Arccore::TraceMessage;
+using Arccore::TraceMessageDbg;
+using Arccore::TraceMessageListenerArgs;
+using Arccore::ITraceMessageListener;
+namespace Trace = ::Arccore::Trace;
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+}
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
